addition of task 15

+++ functions_nested_loops/103-fibonacci.c
@@ -9,11 +9,11 @@
 int main(void)
 {
 	int i = 0;
-	long long  counter = 0LL;
-	long long  n1 = 0LL;
-	long long n2 = 1LL;
-	long long final_fibo = 4000000LL;
-	long long n2aux;
+	long  counter = 0;
+	long n1 = 0;
+	 long n2 = 1;
+	long final_fibo = 4000000;
+	long n2aux;
 
 	while (i < final_fibo)
 	{
@@ -26,6 +26,6 @@ int main(void)
 		n1 = n2aux;
 		i++;
 	}
-	printf("%lld\n", counter);
+	printf("%ld\n", counter);
 	return (0);
 }
